refactor: split train_model.py along Dependency Rule boundary (C-15, C-18, C-08)
Extract pure training logic into training_engine.py (Entity layer):
- make(), _SequenceIndices, _process_sequence(), train(), training_loop()
- Zero framework imports — no views_pipeline_core, no os, no datetime
- Independently testable without ModelPathManager

Slim train_model.py to framework wiring only (38 lines):
- train_model_artifact() — artifact I/O, file system, ModelPathManager
- Backward-compatible re-exports for all existing callers

New tests (TDD):
- C-18: End-to-end training smoke test on 8x8 synthetic data (2 lessons)
- C-08: North-Up flip symmetry assertion (geographic round-trip)
- 3 structural tests: module importable, no pipeline_core imports, exports
