+++ tests/test_volume_handler_hard_gates.py
@@ -362,6 +362,85 @@ def test_gate_slice_time_origin_plus_duration_oob():
 # ─── C-23: extrapolate_time() unit tests ─────────────────────────────────────
 
 
+# ─── C-08: North-Up flip symmetry assertion ──────────────────────────────────
+
+
+def test_gate_flip_symmetry_from_df_to_output():
+    """
+    C-08: The North-Up flip in from_df() and the North-Up flip in
+    _valid_cell_indices() must be symmetric. If a cell is at geographic
+    row R in the input DataFrame, it must appear at geographic row R
+    in the output reconstruction.
+
+    This test creates a gradient pattern (value = row index), runs it
+    through from_df(), then verifies that _valid_cell_indices() recovers
+    the original geographic mapping. Any flip mismatch produces inverted
+    or scrambled coordinates.
+    """
+    H, W = 4, 4
+    cfg = {
+        "time_col": "month_id",
+        "id_col": "priogrid_gid",
+        "spatial_cols": ["row", "col"],
+        "identity_cols": ["month_id", "priogrid_gid"],
+        "features": ["value"],
+        "row_offset": 10,
+        "col_offset": 20,
+        "height": H,
+        "width": W,
+        "transformations": {"identity": ["value"]},
+        "derivations": {},
+    }
+
+    # Build DataFrame with value = global row (geographic truth)
+    rows, cols, values, gids = [], [], [], []
+    for r in range(H):
+        for c in range(W):
+            rows.append(10 + r)
+            cols.append(20 + c)
+            values.append(float(10 + r))  # value = geographic row
+            gids.append(1 + r * W + c)
+
+    df = pd.DataFrame({
+        "month_id": [1] * (H * W),
+        "priogrid_gid": gids,
+        "row": rows,
+        "col": cols,
+        "value": values,
+    })
+
+    handler = VolumeHandler.from_df(df, cfg)
+
+    # Now use _valid_cell_indices to extract the reconstruction mapping.
+    # _valid_cell_indices returns (rows, cols, values) for valid cells.
+    # The geographic row of each output cell must match the value we planted.
+    #
+    # We test the round-trip by checking that the data at each valid cell
+    # has value == its original geographic row.
+    val_idx = handler.channel_map.index("value")
+    gid_idx = handler.channel_map.index("priogrid_gid")
+
+    data = handler.data  # [T=1, H=4, W=4, C]
+    for r_array in range(H):
+        for c_array in range(W):
+            gid = data[0, r_array, c_array, gid_idx]
+            if gid > 0:  # valid cell
+                planted_value = data[0, r_array, c_array, val_idx]
+                # The planted value IS the geographic row.
+                # Recover geographic row from array index:
+                # After North-Up flip, array row 0 = south (highest geo row),
+                # array row H-1 = north (lowest geo row).
+                geo_row = cfg["row_offset"] + (H - 1 - r_array)
+                assert planted_value == geo_row, (
+                    f"Flip symmetry broken at array[{r_array},{c_array}]: "
+                    f"value={planted_value} but geo_row={geo_row}. "
+                    f"North-Up flip in from_df and output path are asymmetric."
+                )
+
+
+# ─── C-23: extrapolate_time() unit tests ─────────────────────────────────────
+
+
 def test_extrapolate_time_shape_preservation():
     """
     C-23: extrapolate_time(steps) must return [steps, H, W, C] with
